v2routers: Use hoodid to identify V1 hoods
All V1 hoods are assigned to IDs 10xxx in the hoods table.
Since no new hoods are added, they are contained in this range.

In addition, the hoods table is forever, so deleted V1 hoods
stay there. This makes the hoodid a perfect WHERE criterion for
the v2routers page, so it does not have to be updated for every
deleted hood.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>