map: Add Polyhood support
This actually implements polyhood support for the MAP.

This is not connected to the earlier commit which provides
database support for polyhoods. This patch will work
independent of the earlier one.

Although the KeyXchange does not provide polyhood data so far,
the Monitoring's implementation can already be put in place and
will take up the data as soon as it's there.

Note that since we only provide an additional layer for the
map, the overall footprint of this change is relatively small.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>

Tested-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>