Replace proprietary Thunderforest tiles with Wikimedia tiles
Thunderforest Maps are proprietary and require an api key.

As I quite like the Wikimedia map style as it has a very low
contrast which makes nodes and connections between them stand out
way more.

The map source parameter is updated as well, so old URLs fall back
to the default style.

Signed-off-by: Fabian Bläse <fabian@blaese.de>