db/router_stats_*: Dilute density of older router statistics
To reduce the amount of router stats, this introduces a new set
of router_stats_* tables that store data in 30-minutes-intervals.

The original tables remain, but the data in those will be kept for
a shorter time.

Thus, we have more dense data for recent stats and less dense data
for older stats. By reducing the density from every 5 to every 30
minutes, we reduce size of the "old" data to a sixth of the initial
amount. This should speed up I/O substantially.

Having two tables per category is superior to trimming the data in
one table: Remove data by e.g. modulo queries is slow and will
cause strong fragmentation of the table. By having two tables, we
introduce same data overlap, but on the other hand effectively
provide table partitioning making individual tables smaller.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>