tilestache: use memcache instead of disk for tile storage
So far, tiles have been stored on disk which, in combination with
bad disk performance, made the map slow and laggy.

This patch switches to memcache, which will remove the dependency
on disk performance for loading cached tiles entirely.

Note that installing and configuring memcached will be required
manually like for the other dependencies.

Signed-off-by: Fabian Bläse <fabian@blaese.de>
Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>