scripts/calcglobalstats: only delete routers from main table
The previous implementation used an INNER JOIN for deletion.
This only works if there is an entry in all connected tables,
leaving a lot of devices undeleted after a certain timespan.

To really get all devices, this changes to just deleting from
the main table. The remnants in other tables can be dealt with
by another function to find orphaned entries later.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>