api/alfredX: handle influxdb in batches
Previously, data was sent to influxdb individually for each router.

This is expected to be inefficient, so let's collect all additions
per retention policy now (the latter is required by influxdb).

The downside is that the whole bunch will fail if one dataset is
broken. However, this only applies to the data after parsing the
XML/JSON. If parsing fails, this will be handled as before.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>