db: add and improve scripts for dealing with influx retention
init_influx.py: This should be run initially when the database is
                set up.

update_influx.py: This should be run when the retention policies
                  are changed.

status_influx.pw: This will print the current retention policies
                  from the database.

For the scripts to work, the dictionary with retention policies
needs to be set up in influconfig.py, like shown in the example
file. Furthermore, the hostname, port, and database need to be
set correctly there, too (but not the user/pw, see below).

Since typically the user for the monitoring itself would not be
an admin, but admin access is required for changing the policies,
the scripts would be started by specifying the proper user name
as argument, e.g.

  init_influx.py username

If auth_enabled = false, no user needs to be specified, and the
scripts may just be run without arguments.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>