statistics.html: Use shorter keys for stats json
Like done recently for the router details page, also reduce the
amount of plain data by using short names for the keys in json
stats. While at it, also remove unused fields and improve the
formatting of the query.

This fixes the traffic graph on the statistics page, which was
broken due to the earlier change for the routers.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>