router.html: only show br-client on top of netif list
While we can have both br-client and br-mesh on top of the netif
list, we can only pre-load the data for one of them. This led to
the situation where br-mesh was shown as selected, but no data was
displayed.

Thus, let's just put the new br-client at that priviledged position,
and having br-mesh deselected by default. This should make the
situation more understandable.

Fixes: bd8d0280f660 ("treewide: add support for br-client")

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>